+++ src/layers/convolution.ts
@@ -1,4 +1,5 @@
 import * as tf from '@tensorflow/tfjs';
+import {SymbolicTensor, Tensor} from '@tensorflow/tfjs';
 import {PaddingMode} from '@tensorflow/tfjs-layers/dist/common';
 import {Layer} from '@tensorflow/tfjs-layers/dist/engine/topology';
 import {ConvLayerConfig} from '@tensorflow/tfjs-layers/dist/layers/convolutional';
@@ -6,6 +7,8 @@ import {onnx} from 'onnx-proto';
 
 import {OnnxNode, WeightInitializer} from '../node';
 import {getNamedAttrs, parseAttrOrDefault} from '../util';
+
+import {Constant} from './core';
 
 export type AutoPad = 'SAME_UPPER'|'SAME_LOWER'|'VALID';
 
@@ -30,6 +33,27 @@ export class Conv extends OnnxNode {
     return parseAttrOrDefault(conf.kernel_shape, []).length || 2;
   }
 
+  getTensorAttr(name: string): Tensor {
+    if (this.model.blobValues !== undefined &&
+        this.model.blobValues.hasOwnProperty(name)) {
+      return this.model.blobValues[name];
+    } else if (
+        this.model.nodes !== undefined &&
+        this.model.nodes.hasOwnProperty(name)) {
+      const node = this.model.nodes[name];
+      if (node.opType == 'Constant') {
+        return Constant.getConstantAttr(node);
+      }
+      throw new Error(`Cannot extract tensor attribute '${
+          name}' from layer other than 'Constant'`);
+    }
+    // TODO if model is not trained, we can use the
+    // this.model.blobShapes to extract tensor shape
+    else {
+      throw new Error(`Cannot find tensor attribute '${name}'`);
+    }
+  }
+
   getTfjsLayerConfig(node: onnx.INodeProto): ConvLayerConfig {
     const conf = getNamedAttrs(node.attribute) as ConvNodeConfig;
     const kernelSize = parseAttrOrDefault(conf.kernel_shape) as number[];
@@ -39,18 +63,16 @@ export class Conv extends OnnxNode {
     const padding = Conv.getTfjsPadding(pads, autoPad);
     const dilationRate = parseAttrOrDefault(conf.dilations, 1);
 
-    const w = node.input[1];
-    const b = node.input[2];
-    const weightShape = this.model.blobShapes[w];
-    const filters = weightShape[0];
-    const kernel = this.model.blobValues[w];
-    const bias = this.model.blobValues[b];
+    const kernel = this.getTensorAttr(node.input[1]);
+    const bias = node.input[2] ? this.getTensorAttr(node.input[2]) : null;
+    const filters = kernel.shape[kernel.shape.length - 1];
 
     return {
       kernelSize: kernelSize, strides: strides, padding: padding,
           dilationRate: dilationRate, filters: filters,
           kernelInitializer: new WeightInitializer(kernel),
-          biasInitializer: new WeightInitializer(bias),
+          useBias: Boolean(bias),
+          biasInitializer: bias ? new WeightInitializer(bias) : undefined,
     }
   }
 
@@ -58,5 +80,9 @@ export class Conv extends OnnxNode {
     const dim = Conv.getConvDim(node);
     const conf = this.getTfjsConfig(node) as ConvLayerConfig;
     return dim == 1 ? tf.layers.conv1d(conf) : tf.layers.conv2d(conf);
+  }
+
+  prepareInput(input?: SymbolicTensor[]) {
+    return input.slice(0, 1);
   }
 }