+++ src/layers/merge.ts
@@ -4,6 +4,7 @@ import {Layer} from '@tensorflow/tfjs-layers/dist/engine/topology';
 import {ConcatenateLayerConfig} from '@tensorflow/tfjs-layers/dist/layers/merge';
 import {onnx} from 'onnx-proto';
 
+import {AddLayer, DivLayer, MulLayer, SubLayer} from '../compat/merge';
 import {OnnxNode} from '../node';
 import {getNamedAttrs, parseAttrOrDefault, parseOnnxAxis} from '../util';
 
@@ -30,23 +31,29 @@ export class Concat extends OnnxNode {
 }
 
 export class Add extends OnnxNode {
-  getTfjsLayerConfig(node: onnx.INodeProto) {
-    return {};
-  }
-
   getTfjsLayer(node: onnx.INodeProto): Layer {
     const conf = this.getTfjsConfig(node);
-    return tf.layers.add(conf)
+    return new AddLayer(conf);
+  }
+}
+
+export class Sub extends OnnxNode {
+  getTfjsLayer(node: onnx.INodeProto): Layer {
+    const conf = this.getTfjsConfig(node);
+    return new SubLayer(conf);
   }
 }
 
 export class Mul extends OnnxNode {
-  getTfjsLayerConfig(node: onnx.INodeProto) {
-    return {};
-  }
-
   getTfjsLayer(node: onnx.INodeProto): Layer {
     const conf = this.getTfjsConfig(node);
-    return tf.layers.multiply(conf)
+    return new MulLayer(conf);
+  }
+}
+
+export class Div extends OnnxNode {
+  getTfjsLayer(node: onnx.INodeProto): Layer {
+    const conf = this.getTfjsConfig(node);
+    return new DivLayer(conf);
   }
 }