+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 * 0.1.0
   * Core framework to parse `Models`, `Nodes` and `Weights`
   * Implementation of common layers
-  * SqueezeNet demo
+  * Add SqueezeNet demo
 
 * 0.1.1
   * Add `Add`, `Mul`, `Flatten` and `Reshape` layers
+
+* 0.1.2
+  * Add `Constant`, `MatMul`, `Div` and `Sub` layers
+  * Fix `Reshape` layer and 2D tensors
+  * Add `topK` in SqueezeNet demo
+  * Add MNIST and Emotion Recognition demos